database save completed with successful reload and replay

+++ src/FiddlerTestRunnerConsole/PersistentFiddlerSession.cs
@@ -26,7 +26,7 @@ namespace FiddlerTestRunnerConsole
 
         public string Data { get; set; }
 
-        public int Len { get; set; }
+        public long Len { get; set; }
 
         public void SetSession(Session session)
         {
